Major Refactor. Added local scheduler

+++ ODLib-Common/src/main/kotlin/pt/up/fc/dcc/hyrax/odlib/interfaces/RemoteODCallback.kt
@@ -3,6 +3,6 @@ package pt.up.fc.dcc.hyrax.odlib.interfaces
 import pt.up.fc.dcc.hyrax.odlib.ODUtils
 
 interface RemoteODCallback{
-    var id: Int
+    var id: Long
     fun onNewResult(resultList: List<ODUtils.ODDetection?>)
 }
